requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
pydantic[email]==2.9.2
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-jose[cryptography]==3.3.0
celery==5.4.0
redis==5.0.8
python-dotenv==1.0.1
python-multipart==0.0.9
transliterate==1.10.2
alembic==1.14.1
markupsafe==3.0.2
mako==1.3.9
httptools==0.6.4
pyopenssl==25.0.0
pyyaml==6.0.2
watchfiles==0.24.0
websockets==15.0
deprecated==1.2.18
limits==4.0.1
packaging==24.2
slowapi==0.1.9
wrapt==1.17.2
tenacity==9.0.0
